pyproject.toml
[project]
name = "port-scaner"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-nmap>=0.7.1",
    "tabulate>=0.10.0",
]
